+++ plugin/chain/main.go
@@ -1,23 +1,27 @@
 package main
 
 import (
-	"github.com/ielab/searchrefiner"
+	"fmt"
 	"github.com/gin-gonic/gin"
-	"net/http"
-	"github.com/hscells/cqr"
-	"github.com/hscells/transmute/pipeline"
-	"github.com/hscells/groove/learning"
-	"log"
 	"github.com/go-errors/errors"
+	"github.com/hscells/cqr"
+	"github.com/hscells/cui2vec"
+	"github.com/hscells/groove"
 	"github.com/hscells/groove/analysis"
 	"github.com/hscells/groove/analysis/preqpp"
-	"fmt"
-	"github.com/hscells/cui2vec"
-	"os"
+	"github.com/hscells/groove/combinator"
+	"github.com/hscells/groove/eval"
+	"github.com/hscells/groove/learning"
 	"github.com/hscells/quickumlsrest"
 	"github.com/hscells/transmute"
-	"github.com/hscells/groove/combinator"
+	"github.com/hscells/transmute/pipeline"
+	"github.com/hscells/trecresults"
+	"github.com/ielab/searchrefiner"
 	"github.com/peterbourgon/diskv"
+	"html/template"
+	"log"
+	"net/http"
+	"os"
 	"sync"
 )
 
@@ -26,9 +30,8 @@ type ChainPlugin struct{}
 var (
 	quickrank string
 	quickumls quickumlsrest.Client
-	vector    cui2vec.CompressedEmbeddings
+	vector    cui2vec.Embeddings
 	mapping   cui2vec.Mapping
-	mu        sync.Mutex
 	queries   = make(map[string]learning.CandidateQuery)
 	chain     = make(map[string][]link)
 	// Cache for the statistics of the query performance predictors.
@@ -38,36 +41,106 @@ var (
 		CacheSizeMax: 4096 * 1024,
 		Compression:  diskv.NewGzipCompression(),
 	})
+
+	models = map[string]string{
+		"precision": "dart_precision.xml",
+		"recall":    "dart_recall.xml",
+		"f1":        "dart_f1.xml",
+	}
+
+	transformationType = []string{
+		"Logical Operator Replacement",
+		"Adjacency Range",
+		"MeSH Explosion",
+		"Field Restrictions",
+		"Adjacency Replacement",
+		"Clause Removal",
+		"cui2vec Expansion",
+		"MeSH Parent",
+	}
+
+	transformationDescriptions = []string{
+		"Modifying the logical operators (AND/OR) of a single clause.",
+		"",
+		"Toggle explosion (tree subsumption) on a single MeSH term.",
+		"Adding or removing fields from a single term.",
+		"",
+		"Query reduction by removal of a single clause from a query.",
+		"Query expansion using CUI word embeddings.",
+		"Transform a single MeSH term by rewriting it as the parent term in the ontology.",
+	}
+
+	workMu  sync.Mutex
+	workMap = make(map[string]workResponse)
 )
 
 type templating struct {
-	Query    learning.CandidateQuery
-	Chain    []link
-	Language string
-	RawQuery string
+	Query       learning.CandidateQuery
+	Chain       []link
+	Language    string
+	RawQuery    string
+	Description template.HTML
+	Evaluation  map[string]float64
+	Error       error
 }
 
 type link struct {
-	NumRet int
-	RelRet int
-	Query  string
+	Evaluation      map[string]float64
+	Transformations map[string]bool
+	Query           string
+	Description     template.HTML
 }
 
-func ret(q cqr.CommonQueryRepresentation, s searchrefiner.Server, u string) (int, int, error) {
-	eq, _ := transmute.CompileCqr2PubMed(q)
-	d, err := s.Entrez.Search(eq, s.Entrez.SearchSize(100000))
+type workRequest struct {
+	user                   string
+	model                  string
+	rawQuery               string
+	lang                   string
+	server                 searchrefiner.Server
+	cq                     learning.CandidateQuery
+	selector               learning.QuickRankQueryCandidateSelector
+	transformations        []learning.Transformation
+	appliedTransformations []string
+}
+
+type workResponse struct {
+	nq         learning.CandidateQuery
+	candidates []learning.CandidateQuery
+	err        error
+	done       bool
+	request    workRequest
+	link       link
+}
+
+func ret(q cqr.CommonQueryRepresentation, s searchrefiner.Server, u string) (map[string]float64, error) {
+	gq := groove.NewPipelineQuery("0", "0", q)
+	t, _, err := combinator.NewLogicalTree(gq, s.Entrez, searchrefiner.QueryCacher)
 	if err != nil {
-		return 0, 0, err
+		log.Println(err)
+		return nil, err
 	}
-	foundRel := 0
-	for _, doc := range d {
-		for _, rel := range s.Settings[u].Relevant {
-			if combinator.Document(doc) == rel {
-				foundRel++
-			}
-		}
+	d := t.Documents(searchrefiner.QueryCacher)
+	qrels := trecresults.NewQrelsFile()
+	qrels.Qrels["0"] = make(trecresults.Qrels, len(s.Settings[u].Relevant))
+	for _, pmid := range s.Settings[u].Relevant {
+		qrels.Qrels["0"][pmid.String()] = &trecresults.Qrel{Topic: "0", Iteration: "0", DocId: pmid.String(), Score: 1}
 	}
-	return len(d), foundRel, nil
+	results := d.Results(gq, "0")
+	return eval.Evaluate(
+		[]eval.Evaluator{
+			eval.PrecisionEvaluator,
+			eval.RecallEvaluator,
+			eval.F1Measure,
+			eval.F05Measure,
+			eval.F3Measure,
+			eval.NumRel,
+			eval.NumRet,
+			eval.NumRelRet,
+		},
+		&results,
+		*qrels,
+		"0",
+	), nil
 }
 
 func initiate() error {
@@ -88,44 +161,169 @@ func initiate() error {
 	}
 
 	log.Println("loading model")
-	v, err := cui2vec.LoadModel(f, true)
+	vector, err = cui2vec.NewPrecomputedEmbeddings(f)
 	if err != nil {
 		return err
 	}
-
-	log.Println("compressing model")
-	vector, err = v.Compress()
-	if err != nil {
-		return err
-	}
-	v = nil
 
 	return nil
 }
 
-func (ChainPlugin) Serve(s searchrefiner.Server, c *gin.Context) {
-	//fmt.Println(vector == nil, mapping == nil)
-	//// Load cui2vec components.
-	//if vector == nil || mapping == nil {
-	//	err := initiate()
-	//	if err != nil {
-	//		// Return a 500 error for now.
-	//		log.Println(errors.New("could not initiate cui2vec"))
-	//		c.Status(http.StatusInternalServerError)
-	//		return
-	//	}
-	//}
-	//fmt.Println(vector == nil, mapping == nil)
+func (w workRequest) start() {
+	go func() {
+		log.Println("recieved work request")
 
-	quickrank = searchrefiner.ServerConfiguration.Config.Options["QuicklearnBinary"].(string)
-	quickumls = quickumlsrest.NewClient(searchrefiner.ServerConfiguration.Config.Options["QuickUMLSURL"].(string))
+		// generate candidates and select the best one.
+		nq, candidates, err := transform(w.cq, w.selector, w.transformations...)
+		if err != nil {
+			log.Println(err)
+			workMap[w.user] = workResponse{
+				err:  err,
+				done: true,
+			}
+			return
+		}
+
+		description := fmt.Sprintf(`
+This query was selected from %d variations. 
+The transformation that was applied to this query was <span class="tooltip label label-rounded c-hand" data-tooltip="%s">%s</span>. 
+The optimisation that was applied was %s.`, len(candidates), transformationDescriptions[nq.TransformationID], transformationType[nq.TransformationID], w.model)
+
+		var q string
+		switch w.lang {
+		case "pubmed":
+			q, _ = transmute.CompileCqr2PubMed(nq.Query)
+		default:
+			q, _ = transmute.CompileCqr2Medline(nq.Query)
+		}
+
+		workMu.Lock()
+		defer workMu.Unlock()
+		queries[w.user] = nq
+		if len(chain[w.user]) == 0 {
+			evaluation, err := ret(w.cq.Query, w.server, w.user)
+			if err != nil {
+				log.Println(err)
+				workMap[w.user] = workResponse{
+					err:  err,
+					done: true,
+				}
+				return
+			}
+
+			chain[w.user] = append(chain[w.user], link{
+				Query:       w.rawQuery,
+				Evaluation:  evaluation,
+				Description: template.HTML("This is the original query that was submitted."),
+			})
+		}
+		evaluation, err := ret(nq.Query, w.server, w.user)
+		if err != nil {
+			log.Println(err)
+			workMap[w.user] = workResponse{
+				err:  err,
+				done: true,
+			}
+			return
+		}
+		// This is the mapping of selected transformations to the actual transformation implementation.
+		transformations := map[string]bool{
+			"clause_removal":     false,
+			"cui2vec_expansion":  false,
+			"mesh_parent":        false,
+			"mesh_explosion":     false,
+			"field_restrictions": false,
+			"logical_operator":   false,
+		}
+		for _, t := range w.appliedTransformations {
+			transformations[t] = true
+		}
+		chain[w.user] = append(chain[w.user], link{
+			Query:           q,
+			Evaluation:      evaluation,
+			Transformations: transformations,
+			Description:     template.HTML(description),
+		})
+		workMap[w.user] = workResponse{
+			nq:         nq,
+			candidates: candidates,
+			err:        err,
+			done:       true,
+			request:    w,
+			link: link{
+				Query:       q,
+				Description: template.HTML(description),
+			},
+		}
+		log.Println("sending work response")
+		return
+	}()
+}
+
+func transform(cq learning.CandidateQuery, selector learning.QuickRankQueryCandidateSelector, selectedTransformations ...learning.Transformation) (learning.CandidateQuery, []learning.CandidateQuery, error) {
+	// Generate variations.
+	// Only the transformations which the user has selected will be used in the variation generation.
+	candidates, err := learning.Variations(
+		cq,
+		searchrefiner.ServerConfiguration.Entrez,
+		analysis.NewDiskMeasurementExecutor(statisticsCache),
+		[]analysis.Measurement{
+			preqpp.RetrievalSize,
+			analysis.BooleanClauses,
+			analysis.BooleanKeywords,
+			analysis.BooleanTruncated,
+			analysis.BooleanFields,
+			analysis.MeshNonExplodedCount,
+			analysis.MeshExplodedCount,
+			analysis.MeshKeywordCount,
+			analysis.MeshAvgDepth,
+			analysis.MeshMaxDepth,
+		},
+		selectedTransformations...,
+	)
+	if err != nil {
+		return cq, nil, err
+	}
+
+	log.Printf("generated %d variations\n", len(candidates))
+
+	// Select the best query.
+	nq, _, err := selector.Select(cq, candidates)
+	if err != nil {
+		return cq, nil, err
+	}
+	log.Println("selected using transformation", nq.TransformationID)
+	return nq, candidates, nil
+}
+
+func (ChainPlugin) Serve(s searchrefiner.Server, c *gin.Context) {
+	// Load cui2vec components.
+	if vector == nil || mapping == nil {
+		err := initiate()
+		if err != nil {
+			err := errors.New("could not initiate cui2vec components")
+			c.HTML(http.StatusInternalServerError, "error.html", searchrefiner.ErrorPage{Error: err.Error(), BackLink: "/plugin/chain"})
+			c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
+	}
+
+	// This is the mapping of selected transformations to the actual transformation implementation.
+	transformations := map[string]learning.Transformation{
+		"clause_removal":     learning.NewClauseRemovalTransformer(),
+		"cui2vec_expansion":  learning.Newcui2vecExpansionTransformer(vector, mapping, quickumls),
+		"mesh_parent":        learning.NewMeshParentTransformer(),
+		"mesh_explosion":     learning.NewMeSHExplosionTransformer(),
+		"field_restrictions": learning.NewFieldRestrictionsTransformer(),
+		"logical_operator":   learning.NewLogicalOperatorTransformer(),
+	}
 
 	// Grab the username of the logged in user.
 	u := s.UserState.Username(c.Request)
 
 	// Create an entry in the query expansion map for the user.
 	if _, ok := queries[u]; !ok {
-		fmt.Println("making new query for user")
+		log.Println("making new query for user")
 		queries[u] = learning.CandidateQuery{}
 		chain[u] = []link{}
 	}
@@ -143,23 +341,21 @@ func (ChainPlugin) Serve(s searchrefiner.Server, c *gin.Context) {
 	if _, ok := c.GetPostForm("clear"); ok {
 		queries[u] = learning.CandidateQuery{}
 		chain[u] = []link{}
-		c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/chain.html"), templating{Query: queries[u], Language: lang}))
+		c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/index.html"), templating{Query: queries[u], Language: lang}))
 		return
 	}
 
-	model, ok := c.GetPostForm("model")
+	var model string
+	model, ok = c.GetPostForm("model")
 	if !ok {
-		// Return a 500 error for now.
-		log.Println(errors.New("no model specified"))
-		c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/chain.html"), templating{Query: queries[u], Language: lang}))
-		return
+		model = "precision"
 	}
 
 	// selector is a quickrank candidate selector configured to only select to a depth of one.
 	selector := learning.NewQuickRankQueryCandidateSelector(
 		quickrank,
 		map[string]interface{}{
-			"model-in":    fmt.Sprintf("plugin/chain/%s", model),
+			"model-in":    fmt.Sprintf("plugin/chain/%s", models[model]),
 			"test-metric": "DCG",
 			"test-cutoff": 1,
 			"scores":      "scores.txt",
@@ -168,7 +364,8 @@ func (ChainPlugin) Serve(s searchrefiner.Server, c *gin.Context) {
 	)
 
 	// Respond to a request to expand a brand new query.
-	if query, ok := c.GetPostForm("query"); ok {
+	var query string
+	if query, ok = c.GetPostForm("query"); ok {
 		// Clear any existing queries.
 		queries[u] = learning.CandidateQuery{}
 		chain[u] = []link{}
@@ -186,102 +383,124 @@ func (ChainPlugin) Serve(s searchrefiner.Server, c *gin.Context) {
 
 		bq, err := compiler.Execute(query)
 		if err != nil {
+			c.HTML(http.StatusInternalServerError, "error.html", searchrefiner.ErrorPage{Error: err.Error(), BackLink: "/plugin/chain"})
 			c.AbortWithError(http.StatusInternalServerError, err)
 			return
 		}
 
 		rep, err := bq.Representation()
 		if err != nil {
+			c.HTML(http.StatusInternalServerError, "error.html", searchrefiner.ErrorPage{Error: err.Error(), BackLink: "/plugin/chain"})
 			c.AbortWithError(http.StatusInternalServerError, err)
 			return
 		}
 
 		q := rep.(cqr.CommonQueryRepresentation)
-		cq = learning.NewCandidateQuery(q, "1", nil)
-		numret, relret, err := ret(q, s, u)
-		if err != nil {
-			log.Println(err)
-			c.Status(http.StatusInternalServerError)
+		cq = learning.NewCandidateQuery(q, "", nil)
+	}
+
+	// Get a list of transformations selected by the user.
+	t, ok := c.GetPostFormArray("transformations")
+	if !ok {
+		// If the user selected none, then just use all of them.
+		t = []string{"clause_removal", "cui2vec_expansion", "mesh_parent", "field_restrictions", "logical_operator"}
+	}
+
+	// Here the transformations are filtered to just the ones that have been selected.
+	selectedTransformations := make([]learning.Transformation, len(t))
+	for i, transformation := range t {
+		if v, ok := transformations[transformation]; ok {
+			selectedTransformations[i] = v
+		} else {
+			err := errors.New(fmt.Sprintf("%s is not a valid transformation", transformation))
+			c.HTML(http.StatusInternalServerError, "error.html", searchrefiner.ErrorPage{Error: err.Error(), BackLink: "/plugin/chain"})
+			c.AbortWithError(http.StatusInternalServerError, err)
 			return
 		}
-		chain[u] = append(chain[u], link{Query: query, NumRet: numret, RelRet: relret})
 	}
 
-	if cq.Query == nil {
-		c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/chain.html"), templating{Query: queries[u], Language: lang}))
-		return
-	}
-
-	fmt.Println(cq.Query, lang)
-
-	// Generate variations.
-	candidates, err := learning.Variations(
-		cq,
-		searchrefiner.ServerConfiguration.Entrez,
-		analysis.NewDiskMeasurementExecutor(statisticsCache),
-		[]analysis.Measurement{
-			preqpp.RetrievalSize,
-			analysis.BooleanClauses,
-			analysis.BooleanKeywords,
-			analysis.BooleanTruncated,
-			analysis.BooleanFields,
-			analysis.MeshNonExplodedCount,
-			analysis.MeshExplodedCount,
-			analysis.MeshKeywordCount,
-			analysis.MeshAvgDepth,
-			analysis.MeshMaxDepth,
-		},
-		learning.NewClauseRemovalTransformer(),
-		//learning.Newcui2vecExpansionTransformer(vector, mapping, quickumls),
-		learning.NewMeshParentTransformer(),
-		learning.NewMeSHExplosionTransformer(),
-		learning.NewFieldRestrictionsTransformer(),
-		learning.NewLogicalOperatorTransformer(),
+	var (
+		nq       learning.CandidateQuery
+		response workResponse
 	)
-	if err != nil && err != combinator.ErrCacheMiss {
-		// Return a 500 error for now.
-		log.Println(err)
-		c.Status(http.StatusInternalServerError)
+
+	if response, ok = workMap[u]; !ok && cq.Query != nil && c.Request.Method == "POST" { // If no work exists, create a job.
+		log.Println("sending work request")
+		workMap[u] = workResponse{
+			done: false,
+		}
+		work := workRequest{
+			user:                   u,
+			server:                 s,
+			cq:                     cq,
+			selector:               selector,
+			transformations:        selectedTransformations,
+			appliedTransformations: t,
+			model:                  model,
+			rawQuery:               query,
+			lang:                   lang,
+		}
+		work.start()
+		c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/queue.html"), nil))
 		return
+	} else { // Otherwise, we can either get the results, or continue to wait until the request is fulfilled.
+		if response.done {
+			log.Println("completed work")
+			if response.err != nil {
+				log.Println(response.err)
+				//c.HTML(http.StatusInternalServerError, "error.html", searchrefiner.ErrorPage{Error: response.err.Error(), BackLink: "/plugin/chain"})
+				c.Render(http.StatusAccepted, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/index.html"), templating{Query: queries[u], Language: lang, Chain: chain[u], Description: chain[u][len(chain[u])-1].Description, RawQuery: chain[u][len(chain[u])-1].Query, Error: response.err}))
+				return
+			}
+		} else if ok && !response.done {
+			log.Println("work has not been completed")
+			c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/queue.html"), nil))
+			return
+		} else {
+			if len(chain[u]) > 0 {
+				log.Println("there is no work but a chain exists")
+				c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/index.html"), templating{Query: queries[u], Language: lang, Chain: chain[u], Description: chain[u][len(chain[u])-1].Description, RawQuery: chain[u][len(chain[u])-1].Query, Error: response.err}))
+			} else {
+				log.Println("there is no work and no query has been submitted")
+				c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/index.html"), templating{Language: lang, Error: response.err}))
+			}
+			return
+		}
 	}
 
-	log.Printf("generated %d variations\n", len(candidates))
+	// Only now can we be sure that there is no more work for this user and we can delete the job.
+	delete(workMap, u)
 
-	// Select the best query.
-	mu.Lock()
-	defer mu.Unlock()
-	nq, _, err := selector.Select(cq, candidates)
-	if err != nil && err != combinator.ErrCacheMiss {
-		// Return a 500 error for now.
-		log.Println(err)
-		c.Status(http.StatusInternalServerError)
-		return
+	// This ensures we only even look at 5 features in the past maximum.
+	// This prevents quickrank from being overloaded with features and crashing.
+	if len(nq.Chain) > 5 {
+		nq.Chain = nq.Chain[1:]
 	}
-	queries[u] = nq
-
-	var q string
-	switch lang {
-	case "pubmed":
-		q, _ = transmute.CompileCqr2PubMed(nq.Query)
-	default:
-		q, _ = transmute.CompileCqr2Medline(nq.Query)
-	}
-
-	numret, relret, err := ret(nq.Query, s, u)
-	if err != nil {
-		log.Println(err)
-		c.Status(http.StatusInternalServerError)
-		return
-	}
-	chain[u] = append(chain[u], link{Query: q, NumRet: numret, RelRet: relret})
 
 	// Respond to a regular request.
-	c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/chain.html"), templating{Query: nq, Language: lang, Chain: chain[u], RawQuery: q}))
+	c.Render(http.StatusOK, searchrefiner.RenderPlugin(searchrefiner.TemplatePlugin("plugin/chain/index.html"), templating{
+		Query:       nq,
+		Language:    lang,
+		Chain:       chain[u],
+		RawQuery:    response.link.Query,
+		Description: template.HTML(response.link.Description),
+		Error:       response.err,
+	}))
 	return
 }
 
 func (ChainPlugin) PermissionType() searchrefiner.PluginPermission {
 	return searchrefiner.PluginUser
+}
+
+func (ChainPlugin) Details() searchrefiner.PluginDetails {
+	return searchrefiner.PluginDetails{
+		Title:       "Query Chain Transformer",
+		Description: "Refine Boolean queries with automatic query transformations.",
+		Author:      "ielab",
+		Version:     "20.Sep.2018",
+		ProjectURL:  "https://ielab.io/searchrefiner/",
+	}
 }
 
 var Chain ChainPlugin