+++ web/static/bigbro.js
@@ -4,10 +4,9 @@ let BigBro = {
         user: "",
         server: "",
         events: ["click", "dblclick", "mousedown", "mouseup",
-            "mouseenter", "mouseover", "mouseout", "wheel",
-            "loadstart", "loadend", "load", "unload",
-            "reset", "submit", "scroll", "resize",
-            "cut", "copy", "paste", "select"
+            "mouseenter", "mouseout", "wheel", "loadstart", "loadend", "load",
+            "unload", "reset", "submit", "scroll", "resize",
+            "cut", "copy", "paste", "select", "keydown", "keyup"
         ],
     },
     // init must be called with the user and the server, and optionally a list of
@@ -23,17 +22,30 @@ let BigBro = {
         }
 
         this.ws = new WebSocket(protocol + this.data.server + "/event");
-
         let self = this;
-        for (let i = 0; i < this.data.events.length; i++) {
-            window.addEventListener(this.data.events[i], function (e) {
-                self.log(e, self.data.events[i]);
-            })
-        }
-        return this;
+
+        this.ws.onopen = function (ev) {
+            for (let i = 0; i < self.data.events.length; i++) {
+                window.addEventListener(self.data.events[i], function (e) {
+                    self.log(e, self.data.events[i]);
+                })
+            }
+            bb.log(ev, "bigbroinit");
+        };
+
+        window.onunload = function (ev) {
+            bb.log(ev, "bigbrodeinit");
+            self.ws.close();
+            console.log("hello");
+        };
+        return this
     },
-    // log
-    log: function (e, method) {
+    // log logs an event with a specified method name (normally the actual event name).
+    log: function (e, method, comment) {
+        if (this.ws.readyState !== 1) {
+            console.warn("bigbro websocket unable to connect");
+            return false;
+        }
         let event = {
             target: e.target.tagName,
             name: e.target.name,
@@ -45,10 +57,7 @@ let BigBro = {
             y: e.y,
             screenWidth: window.innerWidth,
             screenHeight: window.innerHeight,
-            actor: {
-                identifier: this.data.user
-            },
-            comment: ""
+            actor: this.data.user
         };
         if (method === "keydown" || method === "keyup") {
             // Which key was actually pressed?
@@ -62,8 +71,10 @@ let BigBro = {
             // Strength of the wheel rotation.
             event.comment = e.deltaY.toString();
         }
-
-        event.comment = event.comment.replace(/(?:\r\n|\r|\n)/g, "\\n");
+        if (comment != null) {
+            event.comment = comment;
+        }
         this.ws.send(JSON.stringify(event));
+        return true
     }
 };